+++ CHANGES.md
@@ -2,6 +2,8 @@
 
 ### New Features and UX Improvements 
 
+* Representative default variables
+  [#590](https://github.com/CCI-Tools/cate/issues/590)
 * Drag and drop files on Cate's main window
   [#597](https://github.com/CCI-Tools/cate/issues/597)
 * Some (text) areas in the UI should be user-selectable